WIP: Adding new vm script to create/update gcloud machines
