+++ test/e2e/set-up.js
@@ -1,8 +1,24 @@
+const execFile = require('child_process').execFile
 const setupEnv = require('./set-up-env');
 
+const DIR_NAME = './scripts/cd/'
+
+const gcloudEnvScript = DIR_NAME + 'gcloud_vm.sh'
+
 const main = async () => {
-  const env = await setupEnv();
-  console.log("env", env)
+  try {
+    const env = await setupEnv();
+    console.log(env);
+    const child = execFile(gcloudEnvScript, [], { env })
+    child.stdout.on('data', function (data) {
+      console.log(data.toString());
+    });
+    child.stderr.on('data', function (data) {
+      console.log(data.toString());
+    });
+  } catch (e) {
+    console.error('Error', e);
+  }
 }
 
 main();
